Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/geometry/DOMRectList-expected.txt

x

PASS DOMRectList is not [NoInterfaceObject]
PASS DOMRectList is not [LegacyArrayClass]
PASS DOMRectList length
PASS DOMRectList indexed getter
PASS DOMRectList item()
